Add JSON export functionality, parsing and scraping improvements (#59)
Summary of changes:
- Added the `dataToExportJSON()` function, which allows scraped data to
be exported in JSON format (for use in Google Calendar API stuff or
something, ask Samyok)
- `scrapeASemester()` should now work at campuses other than just UMNTC
- Regex now properly reads Saturday and Sunday classes, so they should
now load seamlessly into the calendar as well
- Added support for "additional/outlier meetings", i.e. if your class
has an extra meeting one week (e.g. an exam), it should now be reflected
in the calendar (assuming the meeting is in MyU)
- Refactored and cleaned up data-parsing code a little bit, added
documentation to functions
- Tested scraping and exporting on a few different schedules, fixed a
few bugs, (e.g. code erroring out if the schedule's weird; an issue with
the `.ics` file where all of the courses are present on the first day of
class, even if it's not their correct day)

---------

Co-authored-by: doggu <someoneyoumayknow0@gmail.com>
Co-authored-by: brooklyn204 <brooklyn.cooper.mail@gmail.com>
Co-authored-by: blake-hokanson <114708224+blake-hokanson@users.noreply.github.com>